Create prefix-ping.tsv
This summarizes the individual registry endpoints that can be queried to determine whether a prefix has been registered in a given place. 

Col 1: The fictional prefix 'foobar' is used as a token here (needed because it is followed by additional char cruft in one instance. 
Col 2: The status code that is expected if the prefix does not exist. 
Col 3: The response expected in the event that neither a 400 nor 500 is issued.